requirements.txt
python-telegram-bot==20.7
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
python-dotenv==1.0.0
python-Levenshtein==0.23.0 